Develop T3a-3: the history engine loses its widgets
Three moves into develop/dev_history_gui.c, and the engine ends with ZERO direct
GTK or dt_iop_gui_* calls.

dt_dev_history_gui_update() moves whole -- it attaches missing expanders and
refreshes every module widget after a backend reload; its declaration stays in
dev_history.h because the signature carries no GTK type and its callers are the
ones already driving the engine.

The pending-commit throttle queue moves with it: it exists to coalesce
slider-drag input on the GTK main loop (g_timeout_add, gui_throttle timing), and
its headless timeout is zero -- so dt_dev_add_history_item_real() is GUI input
policy, not engine. The engine core it drains into is now public as
dt_dev_history_commit_item_now(), which is the function everything non-GUI
should have been calling all along.

_check_deleted_instances() stops destroying widgets INSIDE history_mutex as
writer -- the hazard the survey ranked hardest in this file. The engine unlinks
the instance, parks it in dev->alliop (a pipe may still hold it), scrubs the
undo references, and reports it on a removed-list out-parameter;
dt_dev_history_gui_update() destroys the widgets after the lock is released,
inside its own freeze window. Same teardown, opposite side of the lock.

The post-commit presentation block (viewport-size refresh for geometry-changing
modules, enable-toggle sync) becomes the third handler installed by
dt_dev_history_gui_init().

Deliberately left: twelve dt_gui_freeze/dt_dev_get_thumbnail_size gates inside
engine functions (pop/reload/compress/truncate). They are headless-safe as they
stand, and hoisting them into GUI-side wrappers renames call sites across the
tree -- a mechanical follow-up, not this cut's risk profile. The async DB-write
job scheduling likewise stays (control/jobs), pending a decision on who owns
background scheduling.

Release, Debug (-Werror), nofeatures all build.
